+++ supervised_benchmarks/dataset_protocols.py
@@ -1,6 +1,13 @@
+# DataUnit is the only exposed array
+# , which is just a mapping(dict) of Port:numpy.ndarray
+# All calculations will be done in numpy
+
 from __future__ import annotations
+
 from abc import abstractmethod
-from typing import Literal, Protocol, NamedTuple, Mapping, FrozenSet, TypeVar
+from typing import Literal, Protocol, NamedTuple, Mapping, FrozenSet
+
+from numpy.typing import NDArray
 
 from supervised_benchmarks.ports import Port
 from variable_protocols.variables import Variable
@@ -9,25 +16,13 @@ SupportedDatasetNames = Literal['MNIST', 'IRaven']
 DataQuery = Mapping[Port, Variable]
 
 
-class DataContentBound(Protocol):
-    def __getitem__(self, index): ...
+FixedTrain: Literal['FixedTrain'] = 'FixedTrain'
+FixedTest: Literal['FixedTest'] = 'FixedTest'
+FixedValidation: Literal['FixedValidation'] = 'FixedValidation'
+All: Literal['All'] = 'All'
+FixedSubsetType = Literal['FixedTrain', 'FixedTest', 'FixedValidation', 'All']
 
-    def __len__(self): ...
-
-    def __contains__(self, item): ...
-
-    def __iter__(self): ...
-
-
-DataContentCov = TypeVar('DataContentCov', bound=DataContentBound, covariant=True)
-DataContentContra = TypeVar('DataContentContra', bound=DataContentBound, contravariant=True)
-DataContent = TypeVar('DataContent', bound=DataContentBound)
-
-FixedTrain = Literal['FixedTrain']
-FixedTest = Literal['FixedTest']
-FixedValidation = Literal['FixedValidation']
-All = Literal['All']
-FixedSubsetType = Literal[All, FixedTrain, FixedTest, FixedValidation]
+DataUnit = Mapping[Port, NDArray]
 
 
 class Subset(Protocol):
@@ -58,29 +53,13 @@ class SampledSubset(NamedTuple):
         return sum(map(lambda x: x[1] - x[0], self.intervals))
 
 
-class Data(Protocol[DataContentCov]):
-    @property
-    @abstractmethod
-    def port(self) -> Port: ...
-
-    @property
-    @abstractmethod
-    def protocol(self) -> Variable: ...
-
-    @property
-    @abstractmethod
-    def subset(self) -> Subset: ...
-
-    @property
-    @abstractmethod
-    def content(self) -> DataContentCov: ...
+class DataSubset(NamedTuple):
+    protocol: Variable
+    subset: Subset
+    content_map: DataUnit
 
 
-class DataPool(Protocol[DataContentCov]):
-    @property
-    @abstractmethod
-    def port(self) -> Port: ...
-
+class DataPool(Protocol):
     @property
     @abstractmethod
     def src_var(self) -> Variable: ...
@@ -91,12 +70,9 @@ class DataPool(Protocol[DataContentCov]):
 
     @property
     @abstractmethod
-    def fixed_subsets(self) -> Mapping[FixedSubsetType, Data[DataContentCov]]: ...
+    def fixed_subsets(self) -> Mapping[FixedSubsetType, DataSubset]: ...
 
-    def subset(self, subset: Subset) -> Data[DataContentCov]: ...
-
-
-DataPortMap = Mapping[Port, DataPool]
+    def subset(self, subset: Subset) -> DataSubset: ...
 
 
 class DataConfig(Protocol):
@@ -108,12 +84,7 @@ class DataConfig(Protocol):
     @abstractmethod
     def type(self) -> Literal['DataConfig']: ...
 
-    def get_data(self) -> DataPortMap: ...
-
-
-class DataPair(Protocol):
-    output: Data
-    target: Data
+    def get_data(self) -> DataPool: ...
 
 
 class Dataset(Protocol):
@@ -123,6 +94,6 @@ class Dataset(Protocol):
 
     @property
     @abstractmethod
-    def name(self) -> SupportedDatasetNames: ...
+    def name(self) -> str: ...
 
-    def retrieve(self, query: DataQuery) -> Mapping[Port, DataPool]: ...
+    def retrieve(self, query: DataQuery) -> DataPool: ...