+++ supervised_benchmarks/metrics.py
@@ -1,25 +1,22 @@
 import numpy as np
 from numpy.typing import NDArray
-from variable_protocols.base_variables import BaseVariable
-from variable_protocols.protocols import VariableTensor, Variable
-from variable_protocols.variables import var_scalar, ordinal, one_hot
 
 from supervised_benchmarks.metric_protocols import MetricResult, PairMetricImp, PairMetricType
+from variable_protocols.base_variables import BaseVariable
+from variable_protocols.protocols import VariableTensor, Variable
 
 
-# noinspection PyTypeChecker
-# Because Pycharm sucks
 def get_pair_metric(metric_type: PairMetricType, protocol: Variable):
     if metric_type == "mean_acc":
         assert isinstance(protocol, VariableTensor)
         assert isinstance(protocol.var, BaseVariable)
         if protocol.var.type_name == 'ordinal':
-            def mean_acc_numpy(output_class: NDArray, target_class: NDArray):
+            def mean_acc(output: NDArray, target: NDArray) -> MetricResult:
                 return MetricResult(
-                    content=np.mean(output_class == target_class).item(),
+                    content=np.mean(output == target).item(),
                     result_type=metric_type)
         elif protocol.var.type_name == '1hot':
-            def mean_acc_numpy(output: NDArray, target: NDArray):
+            def mean_acc(output: NDArray, target: NDArray) -> MetricResult:
                 target_class = np.argmax(target, axis=1)
                 output_class = np.argmax(output, axis=1)
                 return MetricResult(
@@ -30,20 +27,7 @@ def get_pair_metric(metric_type: PairMetricType, protocol: Variable):
         return PairMetricImp(
             protocol=protocol,
             type=metric_type,
-            measure=mean_acc_numpy
+            measure=mean_acc
         )
     else:
         raise NotImplementedError(f"metric type {metric_type} is not implemented yet")
-
-#
-#     return PairMetricImp(
-#         protocol=var_scalar(one_hot(n_category)),
-#         type='mean_acc',
-#         measure=mean_acc_numpy
-#     )
-
-
-# def accuracy(output: NDArray, target: NDArray) -> float:
-#     target_class = jnp.argmax(target, axis=1)
-#     output_class = jnp.argmax(output, axis=1)
-#     return jnp.mean(output_class == target_class)