+++ supervised_benchmarks/tests/test_dataset_utils.py
@@ -1,0 +1,2 @@
+def test_merge_vec():
+    assert False