+++ supervised_benchmarks/uci_income/uci_income.py
@@ -2,70 +2,50 @@ from __future__ import annotations
 
 from pathlib import Path
 from pprint import pprint
-from typing import NamedTuple, Literal, Mapping, FrozenSet, Dict, TypedDict
+from typing import NamedTuple, Literal, Mapping, FrozenSet, Dict
 
 import numpy.typing as npt
 
-from supervised_benchmarks.dataset_protocols import Subset, DataQuery, Data, DataPortMap, FixedSubset, FixedSubsetType, FixedTrain, FixedTest
-from supervised_benchmarks.ports import Port, Input, Output, Context, OutputOptions, AllVars
-from supervised_benchmarks.uci_income.consts import row_width, TabularDataInfo
+from supervised_benchmarks.dataset_protocols import Subset, DataQuery, DataSubset, FixedSubset, FixedSubsetType, \
+    FixedTrain, FixedTest
+from supervised_benchmarks.ports import Port, Input, Output
+from supervised_benchmarks.uci_income.consts import row_width, TabularDataInfo, get_anynet_feature, AnyNetDiscrete, \
+    AnyNetContinuous, AnyNetDiscreteOut
 from supervised_benchmarks.uci_income.utils import analyze_data, load_data
-from variable_protocols.variables import Variable, ordinal, dim, var_tensor, var_group, gaussian
+from variable_protocols.variables import Variable
 
 name: Literal["UciIncome"] = "UciIncome"
+
+
 # support column names
-
-
-class TabularDataContentNp(TypedDict):
-    symbol_names: tuple[str]
-    value_names: tuple[str]
-    symbols: npt.ArrayLike
-    values: npt.ArrayLike
-
-
-class UciIncomeData(NamedTuple):
-    port: Port
-    protocol: Variable
-    subset: Subset
-    content: TabularDataContentNp
 
 
 class UciIncomeDataPool(NamedTuple):
     data_info: TabularDataInfo
-    array_dict: Mapping[str, npt.ArrayLike]
-    fixed_datasets: dict[FixedSubsetType, Data[TabularDataContentNp]]
+    array_dict: Mapping[str, npt.NDArray]
+    fixed_datasets: Mapping[FixedSubsetType, DataSubset]
     port: Port
     src_var: Variable
     tgt_var: Variable
 
-    def subset(self, subset: Subset) -> Data[TabularDataContentNp]:
+    def subset(self, subset: Subset) -> DataSubset:
         assert self.src_var == self.tgt_var
         raise NotImplementedError
 
-        # noinspection PyTypeChecker
-        # because pyCharm sucks
         # return UciIncomeData(self.port, self.tgt_var, subset, data_array)
-
-
-# noinspection PyTypeChecker
-# because pyCharm sucks
-def get_anynet_feature(_dict_size: int, _n_features: int) -> Variable:
-    return var_group(
-        {var_tensor(gaussian(0, 1), {dim("Feature", _n_features)}),
-         var_tensor(ordinal(_dict_size), {dim("Feature", _n_features)})})
 
 
 dict_size = 10
 
-uci_income_all_anynet = get_anynet_feature(dict_size, row_width)
-uci_income_in_anynet = get_anynet_feature(dict_size, row_width - 1)
-uci_income_out_anynet = get_anynet_feature(dict_size, 1)
+uci_income_in_anynet_discrete = get_anynet_feature(dict_size, row_width - 1, continuous=False)
+uci_income_in_anynet_continuous = get_anynet_feature(dict_size, row_width - 1, continuous=True)
+uci_income_out_anynet_discrete = get_anynet_feature(dict_size, 1, continuous=False)
 
 
 class UciIncome:
     @property
     def ports(self) -> FrozenSet[Port]:
-        return frozenset({Input, Output, AllVars})
+        return frozenset({AnyNetDiscrete, AnyNetContinuous, AnyNetDiscreteOut})
 
     def __init__(self, base_path: Path) -> None:
         # TODO implement download logic
@@ -73,13 +53,12 @@ class UciIncome:
         pprint(self.data_info.common_values)
         pprint(self.data_info.symbol_id_table)
 
-
         # dict_size = len(data_info.symbol_id_table) + 3
 
         symbol_table_tr, value_table_tr = load_data(self.data_info, is_train=True)
         symbol_table_tst, value_table_tst = load_data(self.data_info, is_train=False)
 
-        self.array_dict: Dict[str, npt.ArrayLike] = {
+        self.array_dict: Dict[str, npt.NDArray] = {
             'tr_symbol': symbol_table_tr,
             'tr_value': value_table_tr,
             'tst_symbol': symbol_table_tst,
@@ -88,58 +67,53 @@ class UciIncome:
 
         # noinspection PyTypeChecker
         # because pyCharm sucks
-        self.protocols: Mapping[str, Variable] = {
-            Input: uci_income_in_anynet,
-            Output: uci_income_out_anynet,
-            AllVars: uci_income_all_anynet
+        self.protocols: Mapping[Port, Variable] = {
+            AnyNetDiscrete: uci_income_in_anynet_discrete,
+            AnyNetContinuous: uci_income_in_anynet_continuous,
+            AnyNetDiscreteOut: uci_income_out_anynet_discrete
         }
 
     @property
     def name(self) -> Literal['UciIncome']:
         return name
 
-    def get_fixed_datasets(self, port: Port, variable: Variable) -> Dict[FixedSubsetType, Data[TabularDataContentNp]]:
-        assert port in self.ports
+    def get_fixed_datasets(self, variable: Variable) -> Mapping[FixedSubsetType, DataSubset]:
         n_samples_tr = self.data_info.n_rows_tr
         n_samples_tst = self.data_info.n_rows_tst
-        n_samples = n_samples_tr + n_samples_tst
 
-        # noinspection PyTypeChecker
-        # because pyCharm sucks https://youtrack.jetbrains.com/issue/PY-49439
-        fixed_datasets: Dict[FixedSubsetType, UciIncomeData] = {
-            FixedTrain: UciIncomeData(port,
-                                      variable,
-                                      FixedSubset(FixedTrain, n_samples_tr),
-                                      TabularDataContentNp(
-                                          symbols=self.array_dict[f'tr_symbol'],
-                                          values=self.array_dict[f'tr_value']
-                                      ))
+        fixed_datasets: Dict[FixedSubsetType, DataSubset] = {
+            FixedTrain: DataSubset(variable,
+                                   FixedSubset(FixedTrain, n_samples_tr),
+                                   {
+                                       AnyNetDiscrete: self.array_dict['tr_symbol'][:-1],
+                                       AnyNetContinuous: self.array_dict['tr_value'],
+                                       AnyNetDiscreteOut: self.array_dict['tr_value'][-1, :]
+                                   }),
+            FixedTest: DataSubset(variable,
+                                  FixedSubset(FixedTest, n_samples_tst),
+                                  {
+                                      AnyNetDiscrete: self.array_dict['tr_symbol'][:-1],
+                                      AnyNetContinuous: self.array_dict['tr_value'],
+                                      AnyNetDiscreteOut: self.array_dict['tr_value'][-1, :]
+                                  })
         }
-        # noinspection PyTypeChecker
-        # because pyCharm sucks https://youtrack.jetbrains.com/issue/PY-49439
         return fixed_datasets
 
-    def retrieve(self, query: DataQuery) -> Mapping[Port, UciIncomeDataPool]:
+    def retrieve(self, query: DataQuery) -> UciIncomeDataPool:
         assert all(port in self.ports for port in query)
 
-        return {
-            port: UciIncomeDataPool(
+        return UciIncomeDataPool(
                 array_dict=self.array_dict,
                 data_info=self.data_info,
-                fixed_datasets=self.get_fixed_datasets(port, variable_protocol),
-                port=port,
+                fixed_datasets=self.get_fixed_datasets(variable_protocol),
                 src_var=self.protocols[port],
                 tgt_var=variable_protocol)
-            for port, variable_protocol in query.items()
-        }
 
 
-# noinspection PyTypeChecker
-# Because pycharm sucks
 class UciIncomeDataConfig(NamedTuple):
     base_path: Path
     port_vars: DataQuery
     type: Literal['DataConfig'] = 'DataConfig'
 
-    def get_data(self) -> DataPortMap:
+    def get_data(self) -> UciIncomeDataPool:
         return UciIncome(self.base_path).retrieve(self.port_vars)